# openapi_dump.json
{"openapi":"3.1.0","info":{"title":"Online Course Platform","version":"1.0.0"},"paths":{"/auth/signup":{"post":{"tags":["Authentication"],"summary":"Signup","operationId":"signup_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh Token","operationId":"refresh_token_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/student/enroll":{"post":{"tags":["Student"],"summary":"Enroll Course","operationId":"enroll_course_student_enroll_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrollmentBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrollmentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/student/courses":{"get":{"tags":["Student"],"summary":"Get Courses","operationId":"get_courses_student_courses_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CourseResponse"},"type":"array","title":"Response Get Courses Student Courses Get"}}}}}}},"/student/courses/{category_id}":{"get":{"tags":["Student"],"summary":"Get Courses By Category","operationId":"get_courses_by_category_student_courses__category_id__get","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"title":"Category Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponse"},"title":"Response Get Courses By Category Student Courses  Category Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/student/enrollments/{student_id}":{"get":{"tags":["Student"],"summary":"Get My Enrollments","operationId":"get_my_enrollments_student_enrollments__student_id__get","parameters":[{"name":"student_id","in":"path","required":true,"schema":{"type":"integer","title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EnrollmentResponse"},"title":"Response Get My Enrollments Student Enrollments  Student Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/student/rate":{"post":{"tags":["Student"],"summary":"Rate Course","operationId":"rate_course_student_rate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrollmentRating"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrollmentRatingResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/student/courses/{course_id}/instructor/{instructor_id}/rating":{"get":{"tags":["Student"],"summary":"View Course Rating","operationId":"view_course_rating_student_courses__course_id__instructor__instructor_id__rating_get","parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"integer","title":"Course Id"}},{"name":"instructor_id","in":"path","required":true,"schema":{"type":"integer","title":"Instructor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/student/instructor/{instructor_id}/course/{course_id}/quiz/{quiz_id}/student/{student_id}/attempt":{"post":{"tags":["Student"],"summary":"Attempt Quiz","operationId":"attempt_quiz_student_instructor__instructor_id__course__course_id__quiz__quiz_id__student__student_id__attempt_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"integer","title":"Course Id"}},{"name":"quiz_id","in":"path","required":true,"schema":{"type":"integer","title":"Quiz Id"}},{"name":"student_id","in":"path","required":true,"schema":{"type":"integer","title":"Student Id"}},{"name":"instructor_id","in":"path","required":true,"schema":{"type":"integer","title":"Instructor Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizScoreBase"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/student/quizscore/{course_id}/quiz/{quiz_id}/student/{student_id}":{"get":{"tags":["Student"],"summary":"Get Score By Student","operationId":"get_score_by_student_student_quizscore__course_id__quiz__quiz_id__student__student_id__get","parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"integer","title":"Course Id"}},{"name":"student_id","in":"path","required":true,"schema":{"type":"integer","title":"Student Id"}},{"name":"quiz_id","in":"path","required":true,"schema":{"type":"integer","title":"Quiz Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScoreResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/teacher/category":{"post":{"tags":["Teacher"],"summary":"Create Category","operationId":"create_category_teacher_category_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/teacher/courses":{"post":{"tags":["Teacher"],"summary":"Create Course","operationId":"create_course_teacher_courses_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/teacher/courses/{instructor_id}":{"get":{"tags":["Teacher"],"summary":"Get My Courses","operationId":"get_my_courses_teacher_courses__instructor_id__get","parameters":[{"name":"instructor_id","in":"path","required":true,"schema":{"type":"integer","title":"Instructor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponse"},"title":"Response Get My Courses Teacher Courses  Instructor Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/teacher/{user_id}/courses/{course_id}":{"delete":{"tags":["Teacher"],"summary":"Delete Course","operationId":"delete_course_teacher__user_id__courses__course_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"integer","title":"Course Id"}},{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/teacher/enrollments/{instructor_id}":{"get":{"tags":["Teacher"],"summary":"Get Enrollments","operationId":"get_enrollments_teacher_enrollments__instructor_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"instructor_id","in":"path","required":true,"schema":{"type":"integer","title":"Instructor Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EnrollmentResponse"},"title":"Response Get Enrollments Teacher Enrollments  Instructor Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/teacher/quizzes/instructor/{instructor_id}/{course_id}":{"post":{"tags":["Teacher"],"summary":"Create Quiz","operationId":"create_quiz_teacher_quizzes_instructor__instructor_id___course_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"instructor_id","in":"path","required":true,"schema":{"type":"integer","title":"Instructor Id"}},{"name":"course_id","in":"path","required":true,"schema":{"type":"integer","title":"Course Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizBase"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/teacher/quizzes/instructor/{instructor_id}/{course_id}/public":{"post":{"tags":["Teacher"],"summary":"Create Quiz Public","description":"Public/test endpoint (NO LOGIN REQUIRED).\nMirrors `POST /teacher/quizzes/instructor/{instructor_id}/{course_id}` but skips auth.\n\nNOTE: Do NOT use this in production; it's intended only for local testing.","operationId":"create_quiz_public_teacher_quizzes_instructor__instructor_id___course_id__public_post","parameters":[{"name":"instructor_id","in":"path","required":true,"schema":{"type":"integer","title":"Instructor Id"}},{"name":"course_id","in":"path","required":true,"schema":{"type":"integer","title":"Course Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizBase"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/teacher/lessons/{instructor_id}/{course_id}":{"post":{"tags":["Teacher"],"summary":"Create Lessonn","operationId":"create_lessonn_teacher_lessons__instructor_id___course_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"instructor_id","in":"path","required":true,"schema":{"type":"integer","title":"Instructor Id"}},{"name":"course_id","in":"path","required":true,"schema":{"type":"integer","title":"Course Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonsBase"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Teacher"],"summary":"Get Lesson","operationId":"get_lesson_teacher_lessons__instructor_id___course_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"instructor_id","in":"path","required":true,"schema":{"type":"integer","title":"Instructor Id"}},{"name":"course_id","in":"path","required":true,"schema":{"type":"integer","title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LessonsResponse"},"title":"Response Get Lesson Teacher Lessons  Instructor Id   Course Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/teacher/quizzes/generate":{"post":{"tags":["Teacher"],"summary":"Generate Quiz With Ai","description":"Public/test endpoint (NO LOGIN REQUIRED).\nGenerate quiz using Groq AI based on natural language instructions.\n\nNOTE: If you want the secured version (teacher-only), use:\n`POST /teacher/quizzes/generate/secure`","operationId":"generate_quiz_with_ai_teacher_quizzes_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizGenerationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizGenerationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/teacher/quizzes/generate/secure":{"post":{"tags":["Teacher"],"summary":"Generate Quiz With Ai Secure","description":"Secured endpoint (TEACHER LOGIN REQUIRED).\nSame behavior as `POST /teacher/quizzes/generate` but enforces teacher auth.","operationId":"generate_quiz_with_ai_secure_teacher_quizzes_generate_secure_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizGenerationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizGenerationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/teacher/quizzes/generate/public":{"post":{"tags":["Teacher"],"summary":"Generate Quiz With Ai Public","description":"Backwards-compatible alias for public access.\nPrefer using `POST /teacher/quizzes/generate`.","operationId":"generate_quiz_with_ai_public_teacher_quizzes_generate_public_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizGenerationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizGenerationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/register":{"post":{"tags":["Users"],"summary":"Register User","operationId":"register_user_users_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"tags":["Users"],"summary":"Get User","operationId":"get_user_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users"],"summary":"Delete User","operationId":"delete_user_users__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"CategoryBase":{"properties":{"CategoryName":{"type":"string","title":"Categoryname"}},"type":"object","required":["CategoryName"],"title":"CategoryBase"},"CategoryResponse":{"properties":{"CategoryName":{"type":"string","title":"Categoryname"},"CategoryId":{"type":"integer","title":"Categoryid"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["CategoryName","CategoryId","created_at"],"title":"CategoryResponse"},"CourseBase":{"properties":{"CourseName":{"type":"string","title":"Coursename"},"InstructorId":{"type":"integer","title":"Instructorid"},"Categoryid":{"type":"integer","title":"Categoryid"}},"type":"object","required":["CourseName","InstructorId","Categoryid"],"title":"CourseBase"},"CourseResponse":{"properties":{"CourseName":{"type":"string","title":"Coursename"},"InstructorId":{"type":"integer","title":"Instructorid"},"Categoryid":{"type":"integer","title":"Categoryid"},"courseId":{"type":"integer","title":"Courseid"},"ratings":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Ratings"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["CourseName","InstructorId","Categoryid","courseId","ratings","created_at"],"title":"CourseResponse"},"EnrollmentBase":{"properties":{"CourseId":{"type":"integer","title":"Courseid"},"InstructorId":{"type":"integer","title":"Instructorid"}},"type":"object","required":["CourseId","InstructorId"],"title":"EnrollmentBase"},"EnrollmentRating":{"properties":{"StudentId":{"type":"integer","title":"Studentid"},"CourseId":{"type":"integer","title":"Courseid"},"InstructorId":{"type":"integer","title":"Instructorid"},"rating":{"type":"integer","title":"Rating"}},"type":"object","required":["StudentId","CourseId","InstructorId","rating"],"title":"EnrollmentRating"},"EnrollmentRatingResponse":{"properties":{"Enrollmentnumber":{"type":"integer","title":"Enrollmentnumber"},"StudentId":{"type":"integer","title":"Studentid"},"CourseId":{"type":"integer","title":"Courseid"},"InstructorId":{"type":"integer","title":"Instructorid"},"rating":{"type":"integer","title":"Rating"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["Enrollmentnumber","StudentId","CourseId","InstructorId","rating","created_at"],"title":"EnrollmentRatingResponse"},"EnrollmentResponse":{"properties":{"Enrollmentnumber":{"type":"integer","title":"Enrollmentnumber"},"StudentId":{"type":"integer","title":"Studentid"},"CourseId":{"type":"integer","title":"Courseid"},"InstructorId":{"type":"integer","title":"Instructorid"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["Enrollmentnumber","StudentId","CourseId","InstructorId","created_at"],"title":"EnrollmentResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LessonsBase":{"properties":{"lessonName":{"type":"string","title":"Lessonname"}},"type":"object","required":["lessonName"],"title":"LessonsBase"},"LessonsResponse":{"properties":{"lessonName":{"type":"string","title":"Lessonname"},"lessonnumber":{"type":"integer","title":"Lessonnumber"},"courseId":{"type":"integer","title":"Courseid"},"InstructorId":{"type":"integer","title":"Instructorid"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["lessonName","lessonnumber","courseId","InstructorId","created_at"],"title":"LessonsResponse"},"QuestionData":{"properties":{"QuestionId":{"type":"integer","title":"Questionid"},"QuizId":{"type":"integer","title":"Quizid"},"Questions":{"additionalProperties":true,"type":"object","title":"Questions"}},"type":"object","required":["QuestionId","QuizId","Questions"],"title":"QuestionData"},"QuizBase":{"properties":{"QuizTopic":{"type":"string","title":"Quiztopic"}},"type":"object","required":["QuizTopic"],"title":"QuizBase"},"QuizGenerationRequest":{"properties":{"instructions":{"type":"string","title":"Instructions"}},"type":"object","required":["instructions"],"title":"QuizGenerationRequest"},"QuizGenerationResponse":{"properties":{"quiz":{"$ref":"#/components/schemas/QuizResponse"},"questions":{"$ref":"#/components/schemas/QuestionData"},"question_count":{"type":"integer","title":"Question Count"}},"type":"object","required":["quiz","questions","question_count"],"title":"QuizGenerationResponse"},"QuizResponse":{"properties":{"QuizTopic":{"type":"string","title":"Quiztopic"},"QuizId":{"type":"integer","title":"Quizid"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"InstructorId":{"type":"integer","title":"Instructorid"},"CourseId":{"type":"integer","title":"Courseid"}},"type":"object","required":["QuizTopic","QuizId","created_at","InstructorId","CourseId"],"title":"QuizResponse"},"QuizScoreBase":{"properties":{"QuizScore":{"type":"integer","title":"Quizscore"}},"type":"object","required":["QuizScore"],"title":"QuizScoreBase"},"RefreshTokenRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"RefreshTokenRequest"},"RefreshTokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"RefreshTokenResponse"},"ScoreResponse":{"properties":{"QuizId":{"type":"integer","title":"Quizid"},"StudentId":{"type":"integer","title":"Studentid"},"QuizScore":{"type":"integer","title":"Quizscore"},"CourseId":{"type":"integer","title":"Courseid"},"InstructorId":{"type":"integer","title":"Instructorid"},"attempted_at":{"type":"string","format":"date-time","title":"Attempted At"}},"type":"object","required":["QuizId","StudentId","QuizScore","CourseId","InstructorId","attempted_at"],"title":"ScoreResponse"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","refresh_token","token_type"],"title":"Token"},"UserBase":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"role":{"type":"string","title":"Role"}},"type":"object","required":["name","email","role"],"title":"UserBase"},"UserCreate":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"role":{"type":"string","title":"Role"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","email","role","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}